Apply dhewm3 / 545004a04c8e6c8d5068b8275be5e02d827b5720: Fix ./idlib/math/ for x86_64 sizeof(float) != sizeof(long) on this ABI.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6197 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ idlib/math/Simd_AltiVec.cpp
@@ -59,8 +59,8 @@
 #define IDVEC4_OFFSET 4
 
 // Alignment tests
-#define IS_16BYTE_ALIGNED( x ) ( ( (unsigned long)&x & 0x0F ) == 0 )	
-#define NOT_16BYTE_ALIGNED( x ) ( ( (unsigned long)&x & 0x0F) != 0 )
+#define IS_16BYTE_ALIGNED( x ) ( ( (unsigned int)&x & 0x0F ) == 0 )
+#define NOT_16BYTE_ALIGNED( x ) ( ( (unsigned int)&x & 0x0F) != 0 )
 
 // Aligned storing floats
 #define ALIGNED_STORE2( ADDR, V0, V1 )			\
